resolve reference between users and posts

+++ resolvers/Mutation/post.js
@@ -1,33 +1,46 @@
 const { Posts } = require('../../models');
+const { getUserId } = require('../utils');
 
 const post = {
-  async addPost(_, { title, link, imageUrl }) {
-    const post = await Posts.create({ title, link, imageUrl });
-    return post.id;
+  async addPost(_, { title, link, imageUrl }, context) {
+    const userId = getUserId(context);
+    if (userId) {
+      const post = await Posts.create({ userId, title, link, imageUrl });
+      return post.id;
+    }
+    throw new Error('Not authorized');
   },
 
-  async editPost(_, { id, title, link, imageUrl }) {
-    const [updated] = await Posts.update(
-      { title, link, imageUrl },
-      {
-        where: { id: id }
+  async editPost(_, { id, title, link, imageUrl }, context) {
+    const userId = getUserId(context);
+    if (userId) {
+      const [updated] = await Posts.update(
+        { userId, title, link, imageUrl },
+        {
+          where: { id: id, userId: userId }
+        }
+      );
+      if (updated) {
+        const updatedPost = await Posts.findOne({ where: { id: id } });
+        return updatedPost;
       }
-    );
-    if (updated) {
-      const updatedPost = await Posts.findOne({ where: { id: id } });
-      return updatedPost;
+      throw new Error('Post not updated');
     }
-    return new Error('Post not updated');
+    throw new Error('Not authorized');
   },
 
-  async deletePost(_, { id }) {
-    const deleted = await Posts.destroy({
-      where: { id: id }
-    });
-    if (deleted) {
-      return id;
+  async deletePost(_, { id }, context) {
+    const userId = getUserId(context);
+    if (userId) {
+      const deleted = await Posts.destroy({
+        where: { id: id, userId: userId }
+      });
+      if (deleted) {
+        return id;
+      }
+      throw new Error('Post not deleted');
     }
-    return new Error('Post not deleted');
+    throw new Error('Not authorized');
   },
 }
